ActionDB should use ContentPackResourceMixin
* Always get actions by ref or id in code.
* Deprecate get_action_by_dict and get_action_by_pack_and_name methods.
* Move Access definition from __init__ to base.py
* Update tests.
