+++ st2actions/st2actions/runners/actionchainrunner.py
@@ -27,7 +27,6 @@ from st2common.constants.action import (ACTIONEXEC_STATUS_SUCCEEDED, ACTIONEXEC_
 from st2common.content.loader import MetaLoader
 from st2common.exceptions import actionrunner as runnerexceptions
 from st2common.models.db.action import ActionExecutionDB
-from st2common.models.system.common import ResourceReference
 from st2common.services import action as action_service
 from st2common.util import action_db as action_db_util
 
@@ -195,8 +194,7 @@ class ActionChainRunner(ActionRunner):
             'string': str
         }
 
-        action_db = action_db_util.get_action_by_ref(
-            ResourceReference.from_string_reference(ref=action_ref))
+        action_db = action_db_util.get_action_by_ref(action_ref)
         action_parameters_schema = action_db.parameters
         runnertype_db = action_db_util.get_runnertype_by_name(action_db.runner_type['name'])
         runner_parameters_schema = runnertype_db.runner_parameters
