+++ st2common/st2common/persistence/access.py
@@ -13,10 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from st2common.persistence import Access
+from st2common.exceptions.access import TokenNotFoundError
 from st2common.models.db import MongoDBAccess
 from st2common.models.db.access import UserDB, TokenDB
-from st2common.exceptions.access import TokenNotFoundError
+from st2common.persistence.base import Access
 
 
 class User(Access):
