+++ st2common/st2common/persistence/actionrunner.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from st2common.persistence import Access
+from st2common.persistence.base import Access
 from st2common.models.db.actionrunner import actionrunner_access
