+++ st2common/st2common/persistence/datastore.py
@@ -13,11 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from st2common import persistence
+from st2common.persistence.base import Access
 from st2common.models.db import datastore
 
 
-class KeyValuePair(persistence.Access):
+class KeyValuePair(Access):
     IMPL = datastore.keyvaluepair_access
 
     @classmethod
